Separating the Lexical analysis into a header
By doing so, we now can use lexical as a single function that returns the next token in the code. Instead of the previous behavior, of just running through the entire code storing the tokens in a linked list.
Also started doing some work for the semantic analysis by making the symbols table, but will need to separate it and its functions into a file meant only for the semantic analyzer.